% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_luciole.R
\name{add_luciole}
\alias{add_luciole}
\title{Add Luciole Family to 'sysfonts'}
\usage{
add_luciole(showtext_auto = TRUE)
}
\arguments{
\item{showtext_auto}{Call \code{\link[showtext:showtext_auto]{showtext::showtext_auto()}} to automatically use showtext functionality.}
}
\value{
No value.
}
\description{
Allow to use Luciole font through
package \href{https://cran.r-project.org/package=showtext}{showtext}.
}
\examples{
library(luciole)
library(ggplot2)

# Add font
add_luciole()

# use with ggplot2
ggplot() +
  aes(0, 0, label = "lisibilité") +
  geom_text(size = 20, family = "Luciole") + 
  theme_grey(base_family = "Luciole")

# use with base
plot(0, 0, col = "white")
text(
  x = 0, y = 0,
  labels = "lisibilité", 
  cex = 5,
  family = "Luciole"
)
}
