% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luciole_font.R
\name{luciole-dependency}
\alias{luciole-dependency}
\alias{luciole_font_dependency}
\alias{attach_luciole}
\title{HTML dependency for Luciole typeface}
\usage{
luciole_font_dependency(selector = "body")

attach_luciole(selector = "body")
}
\arguments{
\item{selector}{CSS selector for which to use the font,
usually an HTML tag, default to \code{"body"} (all document).
If \code{NULL} style tag isn't included.}
}
\value{
\itemize{
\item \code{luciole_font_dependency}: an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}.
\item \code{attach_luciole}: an empty \code{div} with Luciole dependency attached.
}
}
\description{
It can be used to integrate Luciole font
into web pages (Shiny applications or RMarkdown documents).
}
\examples{
library(luciole)
library(htmltools)

browsable(tags$div(
  luciole_font_dependency(),
  tag_example(style = "font-size: 3rem;")
))

}
