% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{luciole-paths}
\alias{luciole-paths}
\alias{luciole_path_regular}
\alias{luciole_path_bold}
\alias{luciole_path_italic}
\alias{luciole_path_bolditalic}
\title{Paths to TTF files}
\usage{
luciole_path_regular()

luciole_path_bold()

luciole_path_italic()

luciole_path_bolditalic()
}
\value{
Path to a \code{.ttf} file.
}
\description{
Use to register Luciole for a graphic device
or to copy \code{.ttf} file to another location.
}
\examples{
luciole_path_regular()
}
