% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luciole_font.R
\name{luciole_font}
\alias{luciole_font}
\title{Integrate Luciole typeface in web pages}
\usage{
luciole_font(selector = "body")
}
\arguments{
\item{selector}{CSS selector for which to use the font,
usually an HTML tag, default to \code{"body"} (all document).}
}
\value{
an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}.
}
\description{
Integrate Luciole typeface in web pages
}
\examples{
library(luciole)
library(htmltools)

browsable(tags$div(
  luciole_font(),
  tag_example(),
  tags$h1("First level title"),
  tags$h2("Second level title"),
  tags$h3("Third level title"),
  tags$h4("Fourth level title"),
  tags$h5("Fifth level title"),
  tags$h6("Sixth level title")
))

}
