% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_luciole.R
\name{register_luciole}
\alias{register_luciole}
\title{Register Luciole with systemfonts}
\usage{
register_luciole()
}
\value{
No value.
}
\description{
Register Luciole font to use in your R graphics.
In RStudio, you need to use the \code{AGG} graphics device.
}
\note{
In RStudio, you need to use \code{AGG} as graphics backend to display fonts correctly
to do so, go to
\strong{Tools} > \strong{Global Options} > \strong{General} > \strong{Graphics Device}
and select \code{AGG} in \strong{Backend} menu, then \strong{restart} RStudio.
}
\examples{
library(luciole)
library(ggplot2)

# Register font
register_luciole()

# use with ggplot2
ggplot() +
  aes(0, 0, label = "lisibilité") +
  geom_text(size = 20, family = "Luciole") + 
  theme_grey(base_family = "Luciole")

# use with base
plot(0, 0, col = "white")
text(
  x = 0, y = 0,
  labels = "lisibilité", 
  cex = 5,
  family = "Luciole"
)
}
