% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_example.R
\name{tag_example}
\alias{tag_example}
\title{Character set in HTML tags for examples}
\usage{
tag_example(...)
}
\arguments{
\item{...}{Arguments passed to \code{tags$div}.}
}
\value{
HTML tags.
}
\description{
Character set in HTML tags for examples
}
\examples{
tag_example()
}
